Refactor endpoint definitions into its own file
